Create basic displayImagePreview function
For the HTML, I added a label and an input for the image file uploader.
It accepts multiple files of jpg and png image types. I also added a div
that would contain the image previews of the uploaded images.

For the CSS, I set the opacity of the image preview div to zero so that
it is "hidden" when the images are not yet uploaded.

For the JS, I created a basic logic of the displayImagePreview function.
It is composed of three logic: the first clears the image preview so
that when the user sets another image, it does not stack. The second
dynamically creates the images uploaded if they pass the condition. And,
the third warns the user that they exceed max uploaded images.