Modify HTML structure and style with CSS
I modified the HTML structure of the web app to look like an image file
uploader with accompanying display of the uploaded images.

For the design of the web app, I styled it with a monochromatic color
palette and a custom file input button. The file previews are also
styled to compliment the file uploader.