Revise the image file uploading functionality
I revised the JS script to correctly process and store the images
uploaded by the user. The uploader has input validations to catch errors
in input and displays the proper warning.

For the main logic of the file uploading functionality, I created an
object that holds the file that the user uploaded, and the generated
object URL that references the image of the file. This is then pushed
into an array and is copied into the global image storage.

After that, the uploaded files are showed. For each of the file, a file
preview and an image preview is created and displayed into the sidebar
and main container respectively.

The delete buttons of the file preview works by utilizing the URL
generated and tagged for the file and image preview. When the user
clicks delete, the URL is released, the file and image preview is
removed from DOM, and the global image storage is updated to reflect the
removal of the files.