Restructure project folder for flask backend
I restructed the project folder to prepare for the implementation of the
python flask backend. First, I moved the index.html into the templates
folder. Next, I moved the styles.css and script.js to the static folder.
I made sure that the source and reference of the two are updated in the
html file.

After that I created the app.py file for flask. Then I created a
requirements.txt file to list all dependencies for the project. Finally,
I created a gitignore file to not include the virtual environment from
being pushed into Github.