Create initial flask backend and test form passing
I created the initial flask structure that takes the submitted values
from the form and into the backend. I tested its functionality through
console.logs and terminal logs. It was a success!

I also modified the JS script to fix errors in logic that was captured
through testing. Finally, I updated the requirements.txt file to include
the pillow module.