Add ELWF algorithm and connect to flask backend
I added the Enhanced Local Water-Filling algorithm and its relevant
python files and connected it to the flask backend. I modified the
initial flask backend to fetch the form data from the frontend and pass
it to the processes of the ELWF algorithm. The output of the processes
is then passed back to the frontend for display.

I also updated the script to handle the after processes and modified the
styles to better match the aesthetic of the web app. I also updated the
requirements.txt as dependencies are added.